# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CSM Deployment API",
        "title": "CSM Deployment API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/storage-array-types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all storage array types",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage-array-type"
                ],
                "summary": "List all storage array types",
                "operationId": "list-storage-array-type",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/StorageArrayTypeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/storage-array-types/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a storage array type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage-array-type"
                ],
                "summary": "Get a storage array type",
                "operationId": "get-storage-array-type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Storage Array Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/StorageArrayTypeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/storage-arrays/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get storage array",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage-array"
                ],
                "summary": "Get storage array",
                "operationId": "get-storage-array",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Storage Array ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/StorageArrayResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete storage array",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage-array"
                ],
                "summary": "Delete storage array",
                "operationId": "delete-storage-array",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Storage Array ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a storage array",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage-array"
                ],
                "summary": "Update a storage array",
                "operationId": "update-storage-array",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Storage Array ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Storage Array info for update",
                        "name": "storageArray",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/StorageArrayUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorMessage": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "code": {
                    "description": "HTTPStatusEnum  Possible HTTP status values of completed or failed jobs",
                    "type": "integer",
                    "enum": [
                        200,
                        201,
                        202,
                        204,
                        400,
                        401,
                        403,
                        404,
                        422,
                        429,
                        500,
                        503
                    ]
                },
                "message": {
                    "description": "Message string.",
                    "type": "string"
                },
                "message_l10n": {
                    "description": "Localized message",
                    "type": "object"
                },
                "severity": {
                    "description": "SeverityEnum - The severity of the condition\n* INFO - Information that may be of use in understanding the failure. It is not a problem to fix.\n* WARNING - A condition that isn't a failure, but may be unexpected or a contributing factor. It may be necessary to fix the condition to successfully retry the request.\n* ERROR - An actual failure condition through which the request could not continue.\n* CRITICAL - A failure with significant impact to the system. Normally failed commands roll back and are just ERROR, but this is possible",
                    "type": "string",
                    "enum": [
                        "INFO",
                        "WARNING",
                        "ERROR",
                        "CRITICAL"
                    ]
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "http_status_code": {
                    "description": "HTTPStatusEnum  Possible HTTP status values of completed or failed jobs",
                    "type": "integer",
                    "enum": [
                        200,
                        201,
                        202,
                        204,
                        400,
                        401,
                        403,
                        404,
                        422,
                        429,
                        500,
                        503
                    ]
                },
                "messages": {
                    "description": "A list of messages describing the failure encountered by this request. At least one will\nbe of Error severity because Info and Warning conditions do not cause the request to fail",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ErrorMessage"
                    }
                }
            }
        },
...
        "StorageArrayResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "management_endpoint": {
                    "type": "string"
                },
                "meta_data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "storage_array_type_id": {
                    "type": "string"
                },
                "unique_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "StorageArrayTypeResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "StorageArrayUpdateRequest": {
            "type": "object",
            "properties": {
                "management_endpoint": {
                    "type": "string"
                },
                "meta_data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "password": {
                    "type": "string"
                },
                "storage_array_type": {
                    "type": "string"
                },
                "unique_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}